Don't die if if the Java system property 'os.arch' doesn't exist.
Google App Engine's JVM does not provide the "os.arch" system, but we
shouldn't be failing for this anyways 'cuz argubly the machine type is
Java.  Still, this might be useful for determining JNI libraries to
load.

If 'os.arch' doesn't exist, use 'UNKNOWN' for machine-type.

Thanks to Vladimir V. Korablin for the patch.




git-svn-id: http://abcl.org/svn/trunk/abcl@11750 1c010e3e-69d0-11dd-93a8-456734b0d56f