Implementation of Lisp build without intermediate files.
It is still the case that if anything loaded by 'boot.lisp' causes an
error, the 'abcl.compile.lisp' task does not properly signal an error
to the Ant build process.  But this is no different from the previous
behavior, so we adopt this as the conceptually simpler approach to
writing Ant tasks (and the one that doesn't litter the filesystem with
intermediate build files when the compilation process is being debugged).



git-svn-id: http://abcl.org/svn/trunk@12438 1c010e3e-69d0-11dd-93a8-456734b0d56f