REQUIRE now searches for ASDF systems.
If ASDF is loaded via (REQUIRE 'ASDF), all subsequent invocations of
REQUIRE will search for a loadable ASDF system definitions if the
default resolver mechanism fails.

SYS::*MODULE-PROVIDER-FUNCTIONS* now contains a customizable list of
module provider functions.  Such a function takes a single argument of
the module that should be resolved and loaded.  There is a builtin
resolver #'SYS::MODULE-PROVIDE-SYSTEM that implicitly called before
any functions in this variable.



git-svn-id: http://abcl.org/svn/trunk@12447 1c010e3e-69d0-11dd-93a8-456734b0d56f