Fix for ZipException under interpreted ANSI tests.
Check that the cache entries still accesses an open ZipFile when it is
handed out.

Use SYS:DISABLE-ZIP-CACHE to disable the ZipCache entirely.

Implemented some notion of thread synchronization, although we cannot
guard against the case where two or more references to a ZipFile
exist, and one thread closes the ZipFile.



git-svn-id: http://abcl.org/svn/trunk@12451 1c010e3e-69d0-11dd-93a8-456734b0d56f