Remove UNSAFE-P from SINGLE-VALUED-P (pass2).
Note: The use of UNSAFE-P was misguided. TAGBODY returns NIL, not
any of the values in the body. UNSAFE-P was used to determine (non-local)
returns. BLOCKs do not only return the value of the last form, but can
also return any of the values from the VALUES-FORM in RETURN-FROM. Etc, etc.

git-svn-id: http://abcl.org/svn/branches/unsafe-p-removal@13129 1c010e3e-69d0-11dd-93a8-456734b0d56f