Incorporate output of 'svnversion' into LISP-IMPLEMENTATION-VERSION.
In the Ant-based build, if the executable 'svnversion' can be found on
the PATH and the topmost '.svn' subdirectory exists, the output of
'svnversion' is appended to the value returned by
LISP-IMPLEMENTATION-VERSION.

The use of 'version.src' in 'abcl.properties' has been removed.

TODO: the same functionality should be added to the Lisp-based build.

TODO: test that this works on Windows.

git-svn-id: http://abcl.org/svn/trunk@13197 1c010e3e-69d0-11dd-93a8-456734b0d56f