Reimplement the logic MERGE-PATHNAMES for Pathname version.
I started from scratch interpreting these two passages from the CLHS
Function MERGE-PATHNAMES entry:

"If no version is supplied, default-version is used. If
default-version is nil, the version component will remain unchanged."

"If pathname does not specify a name, then the version, if not
provided, will come from default-pathname, just like the other
components. If pathname does specify a name, then the version is not
affected by default-pathname. If this process leaves the version
missing, the default-version is used."

The previous logic was certainly not obeying the rule where if the
default-version was nil.  Redoing the logic to paraphase the
specification seemed simpler than trying to drawing the corresponding
logic tree and laboriously checking all the cases.

git-svn-id: http://abcl.org/svn/trunk@13291 1c010e3e-69d0-11dd-93a8-456734b0d56f