Fast SHA-{1,256,512} cryptographic hashes for files.
Includes the start of a benchmark that seems to show that using NIO
direcly without involving Lisp streams works a couple orders of
magnitude faster.  A definite point of optimization would be to have a
java.nio.ByteBuffer backed Stream that could be used as source and/or
a sync for fast io via FileChannel.

git-svn-id: http://abcl.org/svn/trunk/abcl@13329 1c010e3e-69d0-11dd-93a8-456734b0d56f