Maven dependencies loaded via ASDF defintions.
If a Maven3 installation can be found via FIND-MVN, ASDF definitions
of the form

   (:mvn "GROUP-ID/ARTIFACT-ID" :version "VERSION")

will be resolved at the ASDF system load time via the Maven Aether
resolution mechanism.

The log4j example wrapping now works.

N.b. this code seriously distorts some of the notions of ASDF which
really seems to depend on its components actually being resident on
the filesystem.  This needs to be adressed ("blessed") via
consultation with ASDF developers to ensure that the path we are
choosing is reasonably future-proof.

git-svn-id: http://abcl.org/svn/trunk/abcl@13550 1c010e3e-69d0-11dd-93a8-456734b0d56f