Fix #177: make logic for finding abcl-contrib more robust.
Issuing a (REQUIRE 'ABCL-CONTRIB) will now use the full name of the
jar archive ABCL was loaded from if it is of the form `abcl.jar' or
`abcl-x.y.z.jar` or `abcl-x.y.z-some-arbitrary-string.jar' to
determine the location of the jar containing the ABCL-CONTRIB
packages.  The namestrings of the ASDF systems located by this
mechanism are now printed to *STANDARD-OUTPUT*.

Installations of the implementations loading from non-standard
locations may use the SYS::*ABCL-JAR* and SYS:*ABCL-CONTRIB* specials
to override this behavior.

git-svn-id: http://abcl.org/svn/trunk/abcl@13730 1c010e3e-69d0-11dd-93a8-456734b0d56f