abcl-asdf: robustify strategy for finding Maven, especially under Windows.
The [Maven installation instructions][maven-install] instruct the user
of Maven to set the environment variables M2_HOME and M2 to reflect
the local installation path, so before we start searching the
environment's PATH, try to probe these locations for the 'mvn'
executable.

[maven-install]: http://maven.apache.org/download.html

Use "where.exe" instead of "which" under Windows for the fallback
strategy if the user has set neither the M2_HOME or M2 variables.
"where.exe" was shipped as part of Windows Server 2003, so it may not
be universally available.

git-svn-id: http://abcl.org/svn/trunk@13844 1c010e3e-69d0-11dd-93a8-456734b0d56f